batch: Move realized entry views into package
The batch implementation has a realization transition in its architecture
checks while realized entry views remain at the package root.

The view model is the base value for provenance-aware content assembly, and
its flat location prevents realization from becoming a bounded dependency.

This commit begins the realization package by moving entry views beneath it
and retaining the selected path during test migration.

Subsequent commits will move provenance, compact storage, boundary lookup,
and mapped range construction.
