+++ src/git_stage_batch/batch/realization/__init__.py
@@ -1,0 +1,1 @@
+"""Realized line entries and their source provenance."""
