+++ src/git_stage_batch/batch/realized_entries.py
@@ -1,20 +1,8 @@
-"""Realized batch entry views."""
+"""Compatibility import for realized entry views."""
 
-from __future__ import annotations
+import sys as _sys
 
-from dataclasses import dataclass
-from typing import Any
+from .realization import entries as _implementation
 
 
-@dataclass(slots=True)
-class RealizedEntry:
-    """A line view in realized content with structural provenance.
-
-    Tracks where each line came from in batch-source space, enabling
-    exact anchored boundary resolution for absence constraints.
-    """
-
-    content: Any  # Line content with newline
-    source_line: int | None  # Batch-source line number, or None for extras
-    target_line: int | None = None  # Working-tree line number, when known
-    is_claimed: bool = False  # True if from a claimed source line
+_sys.modules[__name__] = _implementation
